@@ -1,13 +1,13 @@
-import Typeson from 'typeson';
+import {hasConstructorOf} from 'typeson';
 
 const IntlCollator = {
-    test (x) { return Typeson.hasConstructorOf(x, Intl.Collator); },
+    test (x) { return hasConstructorOf(x, Intl.Collator); },
     replace (c) { return c.resolvedOptions(); },
     revive (options) { return new Intl.Collator(options.locale, options); }
 };
 
 const IntlDateTimeFormat = {
-    test (x) { return Typeson.hasConstructorOf(x, Intl.DateTimeFormat); },
+    test (x) { return hasConstructorOf(x, Intl.DateTimeFormat); },
     replace (dtf) { return dtf.resolvedOptions(); },
     revive (options) {
         return new Intl.DateTimeFormat(options.locale, options);
@@ -15,7 +15,7 @@ const IntlDateTimeFormat = {
 };
 
 const IntlNumberFormat = {
-    test (x) { return Typeson.hasConstructorOf(x, Intl.NumberFormat); },
+    test (x) { return hasConstructorOf(x, Intl.NumberFormat); },
     replace (nf) { return nf.resolvedOptions(); },
     revive (options) { return new Intl.NumberFormat(options.locale, options); }
 };
