@@ -1,13 +1,13 @@
 // This module is for objectified primitives (such as `new Number(3)` or
 //      `new String("foo")`)
 /* eslint-disable no-new-wrappers, unicorn/new-for-builtins */
-import Typeson from 'typeson';
+import {toStringTag} from 'typeson';
 
 const primitiveObjects = {
     // String Object (not primitive string which need no type spec)
     StringObject: {
         test (x) {
-            return Typeson.toStringTag(x) === 'String' && typeof x === 'object';
+            return toStringTag(x) === 'String' && typeof x === 'object';
         },
         replace (s) { return String(s); }, // convert to primitive string
         revive (s) { return new String(s); } // Revive to an objectified string
@@ -15,7 +15,7 @@ const primitiveObjects = {
     // Boolean Object (not primitive boolean which need no type spec)
     BooleanObject: {
         test (x) {
-            return Typeson.toStringTag(x) === 'Boolean' &&
+            return toStringTag(x) === 'Boolean' &&
                 typeof x === 'object';
         },
         replace (b) { return Boolean(b); }, // convert to primitive boolean
@@ -27,7 +27,7 @@ const primitiveObjects = {
     // Number Object (not primitive number which need no type spec)
     NumberObject: {
         test (x) {
-            return Typeson.toStringTag(x) === 'Number' && typeof x === 'object';
+            return toStringTag(x) === 'Number' && typeof x === 'object';
         },
         replace (n) { return Number(n); }, // convert to primitive number
         revive (n) { return new Number(n); } // Revive to an objectified number
