@@ -1,7 +1,7 @@
 /* eslint-env browser, node */
-import Typeson from 'typeson';
+import {toStringTag} from 'typeson';
 
-/* istanbul ignore next */
+/* c8 ignore next */
 const _global = typeof self === 'undefined' ? global : self;
 
 // Support all kinds of typed arrays (views of ArrayBuffers)
@@ -19,12 +19,12 @@ const typedArraysSocketIO = {};
 ].forEach(function (typeName) {
     const arrType = typeName;
     const TypedArray = _global[typeName];
-    /* istanbul ignore if */
+    /* c8 ignore next 3 */
     if (!TypedArray) {
         return;
     }
     typedArraysSocketIO[typeName.toLowerCase()] = {
-        test (x) { return Typeson.toStringTag(x) === arrType; },
+        test (x) { return toStringTag(x) === arrType; },
         replace (a) {
             return (a.byteOffset === 0 &&
                 a.byteLength === a.buffer.byteLength
@@ -42,7 +42,7 @@ const typedArraysSocketIO = {};
             //   get here is an ArrayBuffer
             // If not, let's assume user wants to receive it as
             //   configured with socket.io.
-            return Typeson.toStringTag(buf) === 'ArrayBuffer'
+            return toStringTag(buf) === 'ArrayBuffer'
                 ? new TypedArray(buf)
                 : buf;
         }
