@@ -5,7 +5,7 @@
  */
 function arraybuffer2string (buf) {
     return new Uint8Array(buf).reduce(
-        (s, byte) => s + String.fromCharCode(byte), ''
+        (s, byte) => s + String.fromCodePoint(byte), ''
     );
 }
 
@@ -63,6 +63,8 @@ function string2arraybuffer (str) {
 
     const array = new Uint8Array(str.length);
     for (let i = 0; i < str.length; i++) {
+        // eslint-disable-next-line max-len -- Long
+        // eslint-disable-next-line unicorn/prefer-code-point -- Iterating char. codes
         array[i] = str.charCodeAt(i); // & 0xff;
     }
     return array.buffer;
